в 4 упр радиус делаю float

+++ chapter15/ex04.py
@@ -6,7 +6,7 @@ from math import sqrt
 class Circle:
     """Определяет круг."""
     center: Point
-    radius: int
+    radius: float
 
 
 def create_circle(x: float, y: float, radius: float) -> Circle:
